seller-frontend/pyproject.toml
[project]
name = "seller-frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "grpcio==1.78.0",
    "grpcio-tools==1.78.0",
    "protobuf==6.33.5",
    "typing-extensions==4.15.0",
]
